Исправил замечания и изменил while на do while
